refactor: component 구조 세분화
