+++ front/src/components/auth/SignUpComponent.tsx
