+++ front/src/components/home/HomeComponent.tsx
