+++ front/src/components/index.ts
@@ -1,9 +1,14 @@
-export { default as HomeComponent } from './home/HomeComponents';
+export { default as LoginComponent } from './auth/LoginComponent';
+export { default as SignUpComponent } from './auth/SignUpComponent';
+
+export { default as HomeComponent } from './home/HomeComponent';
+
+export { default as InviteMembBtn } from './invite/InviteMembBtn';
+export { default as InviteMembModal } from './invite/InviteMembModal';
+export { default as InviteChkBtn } from './invite/InviteChkBtn';
+export { default as InviteChkModal } from './invite/InviteChkModal';
 
 export { default as MapComponent } from './map/MapComponent';
-export { default as SearchInput } from './map/SearchInput';
-export { default as SearchResult } from './map/SearchResult';
-export { default as InviteMembBtn } from './map/InviteMembBtn';
-export { default as InviteMembModal } from './map/InviteMembModal';
-export { default as InviteChkBtn } from './map/InviteChkBtn';
-export { default as InviteChkModal } from './map/InviteChkModal';
+
+export { default as SearchInput } from './search/SearchInput';
+export { default as SearchResult } from './search/SearchResult';
