+++ front/src/containers/auth/LoginContainer.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { useNavigate } from 'react-router-dom';
 
 import useInput from 'lib/hooks/useInput';
-import LoginComponents from 'components/auth/LoginComponents';
+import LoginComponents from 'components/auth/LoginComponent';
 import { login } from 'lib/api/auth';
 import { successToast } from '../../utils/toast';
