+++ front/src/lib/api/auth.ts
@@ -1,4 +1,4 @@
-import apiClient from '../apiClient';
+import apiClient from './apiClient';
 
 export const login = (userId: string, userPw: string) => {
     return apiClient.post('/api/auth/login', {
