+++ front/src/lib/api/rest.ts
@@ -1,4 +1,4 @@
-import apiClient from '../apiClient';
+import apiClient from './apiClient';
 
 export const getFavRest = () => {
     return apiClient.get('/api/fav-rest/info', {});
